[Feature/#7] 방 생성 API 구현, 예외처리 메소드 추가, 에러 수정 및 리팩토링

+++ src/main/java/org/prography/spring/dto/responseDto/UserListResponse.java
@@ -12,11 +12,11 @@ import java.util.List;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class UserListResponse {
 
-    private Long totalElements;
-    private Long totalPages;
+    private Integer totalElements;
+    private Integer totalPages;
     private List<UserResponse> users;
 
-    public static UserListResponse of(Long totalElements, Long totalPages, List<UserResponse> users) {
+    public static UserListResponse of(Integer totalElements, Integer totalPages, List<UserResponse> users) {
         return UserListResponse.builder()
                 .totalElements(totalElements)
                 .totalPages(totalPages)
